% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table10.1.R
\docType{data}
\name{table10.1}
\alias{table10.1}
\title{Table 10.1 Chemical Reaction Data}
\format{
A dataframe with 19 rows and 7 columns.
\describe{
\item{ExperimentNumber}{}
\item{y1}{percentage of unchanged starting material}
\item{y2}{percentage converted to the desired product}
\item{y3}{percentage of unwanted by-product}
\item{x1}{temperature}
\item{x2}{concentration}
\item{x3}{time}
}
}
\source{
The data in Table 8.3 were collected by Box and Youle (1955)
}
\usage{
table10.1
}
\description{
The results of a planned experiment involving a chemical reaction are given in Table 10.1
}
\keyword{datasets}
