% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table13.1.R
\docType{data}
\name{table13.1}
\alias{table13.1}
\title{Table 13.1}
\format{
A dataframe with 7 rows and 6 columns.
\describe{
\item{People}{}
\item{Kind}{Adjective}
\item{Intelligent}{Adjective}
\item{Happy}{Adjective}
\item{Likeable}{Adjective}
\item{Just}{Adjective}
}
}
\source{
The data in Table 13.1 from METHODS OF MULTIVARIATE ANALYSIS (Third Edition)
}
\usage{
table13.1
}
\description{
Perception Data: Ratings on Five Adjectives for Seven People
}
\keyword{datasets}
