% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table15.1.R
\docType{data}
\name{table15.1}
\alias{table15.1}
\title{Table 15.1 City Crime Rates per 100,000 Population}
\format{
A dataframe with 16 rows and 8 columns.
\describe{
\item{City}{}
\item{Murder}{A type of crime}
\item{Rape}{A type of crime}
\item{Robbery}{A type of crime}
\item{Assault}{A type of crime}
\item{Burglary}{A type of crime}
\item{Larceny}{A type of crime}
\item{AutoTheft}{A type of crime}
}
}
\source{
The data in Table 15.1 were collected by Hartigan (1975)
}
\usage{
table15.1
}
\description{
Table 15.1 City Crime Rates per 100,000 Population
}
\keyword{datasets}
