% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table15.13.R
\docType{data}
\name{table15.13}
\alias{table15.13}
\title{Table 15.13 Air Pollution Levels in US Cities}
\format{
A dataframe with 41 rows and 8 columns.
\describe{
\item{Cities}{41 US cities}
\item{y1}{SO2 content of air in micrograms per cubic meter}
\item{y2}{Average annual temperature in °F}
\item{y3}{Number of manufacturing enterprises employing 20 or more workers }
\item{y4}{Population size (1970 census) in thousands}
\item{y5}{Average annual wind speed in miles per hour}
\item{y6}{Average annual precipitation in inches}
\item{y7}{Average number of days with precipitation per year}
}
}
\source{
The data in Table 15.8 were collected by Sokal and Rohlf (1981, p. 619)
}
\usage{
table15.13
}
\description{
Table 15.13 Air Pollution Levels in US Cities
}
\keyword{datasets}
