% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table15.14.R
\docType{data}
\name{table15.14}
\alias{table15.14}
\title{Table 15.14 Yields of Winter Wheat (kg per unit area)}
\format{
A dataframe with 12 rows and 5 columns.
\describe{
\item{Site}{twelve different sites in England}
\item{y_1970}{year 1970}
\item{y_1971}{year 1971}
\item{y_1972}{year 1972}
\item{y_1973}{year 1973}
}
}
\source{
The data in Table 15.14 were collected by (Hand et al. 1994, p. 31)
}
\usage{
table15.14
}
\description{
Table 15.14 gives the yields of winter wheat in each of the years 1970-1973 at twelve different sites in England
}
\keyword{datasets}
