% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table15.7.R
\docType{data}
\name{table15.7}
\alias{table15.7}
\title{Table 15.7 Protein Data}
\format{
A dataframe with 25 rows and 10 columns.
\describe{
\item{Country}{}
\item{RedMeat}{}
\item{WhiteMeat}{}
\item{Eggs}{}
\item{Milk}{}
\item{Fish}{}
\item{Cereals}{}
\item{StarchyFoods}{}
\item{Nuts}{}
\item{Fruit/Veg}{}
}
}
\source{
The data in Table 15.7 were collected by Hand et al. (1994, p. 298)
}
\usage{
table15.7
}
\description{
Protein consumption in twenty-five European countries for nine food groups is given in Table 15.7
}
\keyword{datasets}
