% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table16.1.R
\docType{data}
\name{table16.1}
\alias{table16.1}
\title{Table 16.1 Airline Distances Between Ten US Cities}
\format{
A dataframe with 10 rows and 11 columns.
\describe{
\item{City}{}
\item{Atlanta}{}
\item{Chicago}{}
\item{Denver}{}
\item{Houston}{}
\item{LosAngeles}{}
\item{Miami}{}
\item{NewYork}{}
\item{SanFrancisco}{}
\item{Seattle}{}
\item{WashingtonDC}{}
}
}
\source{
The data in Table 15.14 were collected by Kruskal and Wish (1978, pp. 7-9)
}
\usage{
table16.1
}
\description{
Table 16.1 Airline Distances Between Ten US Cities
}
\keyword{datasets}
