% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table16.13.R
\docType{data}
\name{table16.13}
\alias{table16.13}
\title{Table 16.13 Do-It-Yourself Data}
\format{
A dataframe with 24 rows and 7 columns.
\describe{
\item{AccommodationType}{Apartment, House}
\item{Work}{Skilled, Unskilled, Office}
\item{Tenure}{Rent, Own}
\item{Response}{Yes, No}
\item{Age_1}{Up to 30}
\item{Age_2}{31-45}
\item{Age_3}{over 45 }
}
}
\source{
METHODS OF MULTIVARIATE ANALYSIS (Third Edition)
}
\usage{
table16.13
}
\description{
Table 16.13 Do-It-Yourself Data
}
\keyword{datasets}
