% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table16.16.R
\docType{data}
\name{table16.16}
\alias{table16.16}
\title{Table 16.16 Dissimilarity Matrix for World War II Politicians oliticians}
\format{
A dataframe with 12 rows and 13 columns.
\describe{
\item{Person}{}
\item{Hitler}{}
\item{Mussolini}{}
\item{Churchill}{}
\item{Eisenhower}{}
\item{Stalin}{}
\item{Attlee}{}
\item{Franco}{}
\item{DeGaulle}{}
\item{MaoTse}{}
\item{Truman}{}
\item{Chamberlain}{}
\item{Tito}{}
}
}
\source{
Everitt 1987, Table 6.7
}
\usage{
table16.16
}
\description{
Two subjects assessed the degree of dissimilarity between World War II politicians.
}
\keyword{datasets}
