% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table16.17.R
\docType{data}
\name{table16.17}
\alias{table16.17}
\title{Table 16.17}
\format{
A dataframe with 12 rows and 13 columns
\describe{
\item{Birth_Death}{Birth/Death}
\item{Jan}{month}
\item{Feb}{month}
\item{Ma}{month}
\item{Apr}{month}
\item{May}{month}
\item{Jun}{month}
\item{Jul}{month}
\item{Agu}{month}
\item{Sep}{month}
\item{Oct}{month}
\item{Nov}{month}
\item{Dec}{month}
}
}
\source{
Andrews and Herzberg (1985), Table 71.2
}
\usage{
table16.17
}
\description{
Table 16.17 Birth and Death Months of 1281 People
}
\keyword{datasets}
