% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table16.19.R
\docType{data}
\name{table16.19}
\alias{table16.19}
\title{Table 16.19 Byssinosis Data}
\format{
A dataframe with 48 rows and 8 columns.
\describe{
\item{Race}{Other, White}
\item{Smoking}{Non-smoker, Smoker}
\item{Gender}{Female, Male}
\item{Years_in_Job}{}
\item{SufferByssi_nosis}{}
\item{High_dust}{}
\item{Low_dust}{}
\item{Med_dust}{}
}
}
\source{
Andrews and Herzberg (1985, Table 34.1)
}
\usage{
table16.19
}
\description{
Table 16.19 Byssinosis Data
}
\keyword{datasets}
