% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table16.8.R
\docType{data}
\name{table16.8}
\alias{table16.8}
\title{Table 16.8 A List of 12 People and Their Categories on Four Variable}
\format{
A dataframe with 12 rows and 5 columns.
\describe{
\item{Person}{12 People}
\item{Gender}{Male, female }
\item{Age}{Young, middle aged, old}
\item{MaritalStatus}{Single, married }
\item{HairColor}{Blond, brown, black, red}
}
}
\source{
METHODS OF MULTIVARIATE ANALYSIS (Third Edition)
}
\usage{
table16.8
}
\description{
Table 16.8 A List of 12 People and Their Categories on Four Variable
}
\keyword{datasets}
