% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table3.1.R
\docType{data}
\name{table3.1}
\alias{table3.1}
\title{Table 3.1}
\format{
A dataframe with 20 rows and 3 columns.
\describe{
\item{person}{}
\item{x }{}
\item{y }{}
}
}
\source{
Extracted from Table 3.1 in Rencher (3rd ed.)
}
\usage{
table3.1
}
\description{
Table 3.1 Height and Weight for a Sample of 20 College-Age Males
}
\keyword{datasets}
