% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table3.2.R
\docType{data}
\name{table3.2}
\alias{table3.2}
\title{Table 3.2}
\format{
A dataframe with 6 rows and 6 columns.
\describe{
\item{State}{}
\item{y_1932 }{}
\item{y_1936 }{}
\item{y_1940 }{}
\item{y_1960 }{}
\item{y_1964 }{}
\item{y_1968 }{}
}
}
\source{
The data in Table 3.2 are from Kleiner and Hartigan (1981)
}
\usage{
table3.2
}
\description{
Table 3.2 Percentage of Republican Votes in Presidential Elections in Six Southern States for Selected Years
}
\keyword{datasets}
