% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table3.3.R
\docType{data}
\name{table3.3}
\alias{table3.3}
\title{Table 3.4}
\format{
A dataframe with 10 rows and 4 columns.
\describe{
\item{LocationNumber}{}
\item{y1 }{}
\item{y2 }{}
\item{y3 }{}
}
}
\source{
The data set in Table 3.3 contains yield and soil quality measurements at each of 215 locations in a 16-hectare field. The Baker field (Colvin et al., 1997)
}
\usage{
table3.3
}
\description{
Table 3.4 Baker Corn Field Measurements of Yield and Soil Richness
}
\keyword{datasets}
