% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table3.4.R
\docType{data}
\name{table3.4}
\alias{table3.4}
\title{Table 3.4}
\format{
A dataframe with 10 rows and 4 columns.
\describe{
\item{LocationNumber}{}
\item{y1 }{}
\item{y2 }{}
\item{y3 }{}
}
}
\source{
The data set in Table 3.4 Calcium in Soil and Turnip Greens.
}
\usage{
table3.4
}
\description{
Table 3.4 Calcium in Soil and Turnip Greens
}
\keyword{datasets}
