% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table3.5.R
\docType{data}
\name{table3.5}
\alias{table3.5}
\title{Table 3.5}
\format{
A dataframe with 46 rows and 6 columns
\describe{
\item{PatientNumber}{}
\item{y1}{}
\item{y2}{}
\item{x1}{}
\item{x2}{}
\item{x3}{}
}
}
\source{
Book
}
\usage{
table3.5
}
\description{
Table 3.5 Relative Weight, Blood Glucose, and Insulin Levels
}
\keyword{datasets}
