% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table3.6.R
\docType{data}
\name{table3.6}
\alias{table3.6}
\title{Table 3.6}
\format{
A dataframe with 27 rows and 6 columns.
\describe{
\item{SubjectNumber}{ID}
\item{y1}{The variables are response times for the j_th probe word, y_j,j = 1,2,..., 5}
\item{y2}{The variables are response times for the j_th probe word, y_j,j = 1,2,..., 5}
\item{y3}{The variables are response times for the j_th probe word, y_j,j = 1,2,..., 5}
\item{y4}{The variables are response times for the j_th probe word, y_j,j = 1,2,..., 5}
\item{y5}{The variables are response times for the j_th probe word, y_j,j = 1,2,..., 5}
}
}
\source{
Timm (1975, p. 233; 1980, p. 47)
}
\usage{
table3.6
}
\description{
Table 3.6 Response Times for Five Probe Word Positions
}
\keyword{datasets}
