% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table3.7.R
\docType{data}
\name{table3.7}
\alias{table3.7}
\title{Table 3.7}
\format{
A dataframe with 20 rows and 5 columns.
\describe{
\item{Individual}{}
\item{y1 }{}
\item{y2 }{}
\item{y3 }{}
\item{y4 }{}
}
}
\source{
The data in Table 3.7 (Elston and Grizzle 1962) consist of measurements 2/i 12/27 2/3- an d 2/4 °ft n e ramus bone at four different ages on each of 20 boys
}
\usage{
table3.7
}
\description{
Table 3.7 Ramus Bone Length at Four Ages for 20 Boys
}
\keyword{datasets}
