% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table3.8.R
\docType{data}
\name{table3.8}
\alias{table3.8}
\title{Table 3.8}
\format{
A dataframe with 25 rows and 5 columns
\describe{
\item{Group}{First, Second}
\item{y1}{Head Length}
\item{y2}{Head Breath}
\item{x1}{Head Length}
\item{x2}{Head Breath}
}
}
\source{
Frets (1921)
}
\usage{
table3.8
}
\description{
Table 3.8 Measurements on the First and Second Adult Sons in a Sample of 25 Families
}
\keyword{datasets}
