% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table3.9.R
\docType{data}
\name{table3.9}
\alias{table3.9}
\title{Table 3.9 Blood Glucose Measurements on Three Occasions}
\format{
A data frame with 50 rows and 6 variables:
\describe{
\item{y1}{Blood glucose measurement at occasion 1}
\item{y2}{Blood glucose measurement at occasion 2}
\item{y3}{Blood glucose measurement at occasion 3}
\item{x1}{First explanatory variable}
\item{x2}{Second explanatory variable}
\item{x3}{Third explanatory variable}
}
}
\source{
Rencher, Methods of Multivariate Analysis
}
\description{
Blood glucose measurements taken on three different occasions.
}
