% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table5.1.R
\docType{data}
\name{table5.1}
\alias{table5.1}
\title{Table 5.1}
\format{
A dataframe with 64 rows and 5 columns.
\describe{
\item{Group}{}
\item{y1 }{}
\item{y2 }{}
\item{y3 }{}
\item{y4 }{}
}
}
\source{
Four psychological tests were given to 32 men and 32 women. The data are recorded in Table 5.1 (Beall 1945)
}
\usage{
table5.1
}
\description{
Table 5.1 Four Psychological Test Scores on 32 Males and 32 Females
}
\keyword{datasets}
