% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table5.3.R
\docType{data}
\name{table5.3}
\alias{table5.3}
\title{Table 5.3}
\format{
A dataframe with 15 rows and 7 columns.
\describe{
\item{Location}{}
\item{y1 }{}
\item{y2 }{}
\item{x1 }{}
\item{x2 }{}
\item{d1 }{}
\item{d2 }{}
}
}
\source{
Extracted from Table 3.1 in Rencher (3rd ed.)
}
\usage{
table5.3
}
\description{
Table 5.3 Maximum Depth of Pits and Number of Pits of Coated Pipes
}
\keyword{datasets}
