% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table5.5.R
\docType{data}
\name{table5.5}
\alias{table5.5}
\title{Table 5.5}
\format{
A dataframe with 39 rows and 6 columns.
\describe{
\item{Number}{}
\item{Group }{}
\item{y1 }{}
\item{y2 }{}
\item{y3 }{}
\item{y4 }{}
}
}
\source{
Extracted from Table 3.1 in Rencher (3rd ed.)
}
\usage{
table5.5
}
\description{
Table 5.5 Four Measurements on Two Species of Flea Beetles
}
\keyword{datasets}
