% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table5.6.R
\docType{data}
\name{table5.6}
\alias{table5.6}
\title{Table 5.6}
\format{
A dataframe with 20 rows and 12 columns.
\describe{
\item{E_y1 }{}
\item{E_y2 }{}
\item{E_y3 }{}
\item{E_y3 }{}
\item{E_y4 }{}
\item{E_y5 }{}
\item{P_y1 }{}
\item{P_y2 }{}
\item{P_y3 }{}
\item{P_y4 }{}
\item{P_y5 }{}
\item{P_y6 }{}
}
}
\source{
Extracted from Table 3.1 in Rencher (3rd ed.)
}
\usage{
table5.6
}
\description{
Table 5.6 Comparison of Six Tests on Engineer Apprentices and Pilots
}
\keyword{datasets}
