% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table5.7.R
\docType{data}
\name{table5.7}
\alias{table5.7}
\title{Table 5.7 Comparison of Carriers and Noncarriers of Muscular Dystrophy}
\format{
A data frame with 73 rows and 7 variables
\describe{
\item{Group}{Group identifier: Carrier or Noncarrier}
\item{y1}{Biomarker 1 (e.g., enzyme level)}
\item{y2}{Biomarker 2}
\item{y3}{Biomarker 3}
\item{y4}{Biomarker 4}
\item{y5}{Biomarker 5}
\item{y6}{Biomarker 6}
}
}
\source{
Andrews and Herzberg (1985), pp. 222–228. Data were collected in an attempt to find a screening procedure to detect carriers of Duchenne muscular dystrophy, a disease transmitted from female carriers to some of their male offspring
}
\usage{
table5.7
}
\description{
Data from a study comparing carriers and noncarriers of Duchenne muscular dystrophy
}
\keyword{datasets}
