% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table5.8.R
\docType{data}
\name{table5.8}
\alias{table5.8}
\title{Table 5.8}
\format{
A dataframe with 19 rows and 6 columns.
\describe{
\item{Item}{}
\item{Group }{}
\item{y1 }{}
\item{y2 }{}
\item{y3 }{}
\item{y4 }{}
}
}
\source{
Various aspects of economic cycles were measured for consumer goods and producer goods by Tintner (1946)
}
\usage{
table5.8
}
\description{
Table 5.8 Cyclical Measurements of Consumer Goods and Producer Goods
}
\keyword{datasets}
