% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table5.9.R
\docType{data}
\name{table5.9}
\alias{table5.9}
\title{Table5.9}
\format{
A dataframe with 15 rows and 7 columns.
\describe{
\item{Student}{}
\item{y1 }{}
\item{y2 }{}
\item{x1 }{}
\item{x2 }{}
\item{d1 }{}
\item{d2 }{}
}
}
\source{
Each of 15 students wrote an informal and a formal essay Kramer (1972, p. 100)
}
\usage{
table5.9
}
\description{
Table 5.9 Number of Words and Number of Verbs
}
\keyword{datasets}
