% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table6.18.R
\docType{data}
\name{table6.18}
\alias{table6.18}
\title{Table 6.18}
\format{
A dataframe with 60 rows and 7 columns.
\describe{
\item{S}{}
\item{V }{}
\item{ID }{}
\item{y1 }{}
\item{y2 }{}
\item{y3 }{}
\item{y4 }{}
}
}
\source{
Table 6.18 from Keuls et al. (1984)
}
\usage{
table6.18
}
\description{
Table 6.18 Snap Bean Data
}
\keyword{datasets}
