% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table6.19.R
\docType{data}
\name{table6.19}
\alias{table6.19}
\title{Table 6.19}
\format{
A dataframe with 20 rows and 13 columns.
\describe{
\item{Subject}{}
\item{R1_y1 }{}
\item{R1-y2 }{}
\item{R1_y3 }{}
\item{R2_y1 }{}
\item{R2_y2 }{}
\item{R2_y3 }{}
\item{R3_y1 }{}
\item{R3_y2 }{}
\item{R3_y3 }{}
}
}
\source{
In Table 6.19, we have a comparison of four reagents (Burdick 1979)
}
\usage{
table6.19
}
\description{
Table 6.19 Blood Data
}
\keyword{datasets}
