% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table6.23.R
\docType{data}
\name{table6.23}
\alias{table6.23}
\title{Table 6.23}
\format{
A dataframe with 13 rows and 8 columns
\describe{
\item{Mouse}{variable 1}
\item{Day3}{variable 2}
\item{Day6}{variable 3}
\item{Day9}{variable 4}
\item{Day12}{variable 5}
\item{Day15}{variable 6}
\item{Day18}{variable 7}
\item{Day21}{variable 8}
}
}
\source{
Table 6.23 contains the weights of 13 male mice measured every 3 days from birth to weaning. The data set was reported and analyzed by Williams and Izenman (1981) and by Izenman and Williams (1989) and has been further analyzed by Rao (1984, 1987) and by Lee (1988). Analyze as a one-sample growth curve design
}
\usage{
table6.23
}
\description{
Table 6.23 Weights of 13 Male Mice Measured at Successive Intervals of 3 Days over 21 Days from Birth to Weaning
}
\keyword{datasets}
