% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table6.24.R
\docType{data}
\name{table6.24}
\alias{table6.24}
\title{Table 6.24}
\format{
A dataframe with 12 rows and 5 columns.
\describe{
\item{Group}{three groups of trout}
\item{Time_1 }{Time Point}
\item{Time_2 }{Time Point}
\item{Time_3 }{Time Point}
\item{Time_4 }{Time Point}
}
}
\source{
The data set was reported by Beauchamp and Hoel (1973)
}
\usage{
table6.24
}
\description{
In Table 6.24, we have measurements of proportions of albumin at four time points on three groups of trout
}
\keyword{datasets}
