% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table6.25.R
\docType{data}
\name{table6.25}
\alias{table6.25}
\title{Weekly Gains in Weight for 27 Rats}
\format{
A dataframe with 13 rows and 8 columns.
\describe{
\item{Group}{The groups are 1 = controls, 2 = thyroxin added to drinking water, and 3 = thiouracil added to drinking water.}
\item{Rat}{}
\item{y1 }{ gain in week 1}
\item{y2 }{ gain in week 2}
\item{y3 }{ gain in  week 3}
\item{y4 }{ gain in  week 4}
}
}
\source{
The data set was reported by Box (1950)
}
\usage{
table6.25
}
\description{
Table 6.25 contains weight gains for three groups of rats
}
\keyword{datasets}
