% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table6.27.R
\docType{data}
\name{table6.27}
\alias{table6.27}
\title{Blood Pressure Data}
\format{
A dataframe with 31 rows and 7 columns.
\describe{
\item{Group}{group 1 is the controls and groups 2-4 have been exposed to halothane concentrations of .25\%, .50\%, 1.0\% }
\item{M_1}{Number of Minutes after Ligation}
\item{M_5}{Number of Minutes after Ligation}
\item{M_10}{Number of Minutes after Ligation}
\item{M_15}{Number of Minutes after Ligation}
\item{M_30}{Number of Minutes after Ligation}
\item{M_60}{Number of Minutes after Ligation}
}
}
\source{
The data set was reported and by Crepeau et al (1985).
}
\usage{
table6.27
}
\description{
Table 6.27 contains blood pressure measurements at intervals after inducing a heart attack for four groups of rats
}
\keyword{datasets}
