% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table6.28.R
\docType{data}
\name{table6.28}
\alias{table6.28}
\title{Plasma Inorganic Phosphate (mg/dl)}
\format{
A dataframe with 33 rows and 10 columns.
\describe{
\item{Group}{Two control and obese groups }
\item{M_1}{Hours after Glucose Challenge in minutes }
\item{M_5 }{Hours after Glucose Challenge in minutes}
\item{M_10 }{Hours after Glucose Challenge in minutes}
\item{M_15 }{Hours after Glucose Challenge in minutes}
\item{M_30}{Hours after Glucose Challenge in minutes}
\item{M_60 }{Hours after Glucose Challenge in minutes}
}
}
\source{
The data set was reported and by Zerbe (1979)
}
\usage{
table6.28
}
\description{
Table 6.28 compares 13 control and 20 obese patients on a glucose tolerance test using plasma inorganic phosphate
}
\keyword{datasets}
