% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table6.8.R
\docType{data}
\name{table6.8}
\alias{table6.8}
\title{Table 6.8
Table 6.8 Weight of Guinea Pigs Under 3 Levels of Vitamin E Supplements}
\format{
A dataframe with 15 rows and 9 columns
\describe{
\item{Group}{variable 1}
\item{Animal}{variable 1}
\item{Week_1}{variable 1}
\item{Week_2}{variable 1}
\item{Week_3}{variable 1}
\item{Week_4}{variable 1}
\item{Week_5}{variable 1}
\item{Week_6}{variable 1}
\item{Week_7}{variable 1}
}
}
\source{
Three vitamin E diet supplements with levels zero, low, and high were compared for their effect on growth of guinea pigs (Crowder and Hand 1990, pp. 21-29). Five guinea pigs received each supplement level, and their weights were recorded at the end of weeks 1, 3,4, 5, 6, and 7. These weights are given in Table 6.8
}
\usage{
table6.8
}
\description{
Table 6.8
Table 6.8 Weight of Guinea Pigs Under 3 Levels of Vitamin E Supplements
}
\keyword{datasets}
