% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table7.1.R
\docType{data}
\name{table7.1}
\alias{table7.1}
\title{Seishu Measurements}
\format{
A dataframe with 30 rows and 10 columns.
\describe{
\item{y1}{taste}
\item{y2}{odor}
\item{x1}{PH}
\item{x2}{acidity 1}
\item{x3}{acidity 2 }
\item{x4}{ sake meter}
\item{x5}{direct reducing sugar}
\item{x6}{total sugar}
\item{x7}{alcohol}
\item{x8}{formyl-nitrogen}
}
}
\source{
The data set was reported and by Siotani et al. (1963)
}
\usage{
table7.1
}
\description{
Table 7.1 Seishu Measurements
}
\keyword{datasets}
