% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table7.2.R
\docType{data}
\name{table7.2}
\alias{table7.2}
\title{Table 7.2 Temperati, Humidity, and Evaporation}
\format{
A dataframe with 46 rows and 11 columns.
\describe{
\item{y1}{maximum daily air temperature}
\item{y2}{minimum daily air temperature}
\item{y3}{integrated area under daily air temperature curve, that is, a measure of average air temperature}
\item{y4}{maximum daily soil temperature}
\item{y5}{minimum daily soil temperature}
\item{y6}{integrated area under soil temperature curve}
\item{y7}{maximum daily relative humidity}
\item{y8}{minimum daily relative humidity}
\item{y9}{integrated area under daily humidity curve}
\item{y10}{total wind, measured in miles per day}
\item{y11}{evaporation}
}
}
\source{
courtesy of R. J. Freund
}
\usage{
table7.2
}
\description{
The data in Table 7.2 relate temperature, humidity, and evaporation
}
\keyword{datasets}
