% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table8.1.R
\docType{data}
\name{table8.1}
\alias{table8.1}
\title{Table 8.1}
\format{
A dataframe with 12 rows and 3 columns.
\describe{
\item{Temperatures}{maximum daily air temperature}
\item{y1}{yield point}
\item{y2}{ultimate strength}
}
}
\source{
Kramer and Jensen (1969)
}
\usage{
table8.1
}
\description{
Samples of steel produced at two different rolling temperatures are compared in Table 8.1
}
\keyword{datasets}
