% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table8.3.R
\docType{data}
\name{table8.3}
\alias{table8.3}
\title{Table 8.3 Head Measurements for Three Groups}
\format{
A dataframe with 90 rows and 7 columns.
\describe{
\item{Group}{ high school football players (group 1), college football players (group 2), and nonfootball players (group 3)}
\item{WDIM}{head width at widest dimension}
\item{CIRCUM}{head circumference}
\item{FBEYE}{front-to-back measurement at eye level}
\item{EYEHD}{eye-to-top-of-head measurement}
\item{EARHD}{ear-to-top-of-head measurement}
\item{JAW}{jaw width}
}
}
\source{
The data in Table 8.3 were collected by G. R. Bryce and R. M. Barker(Brigham Young University)
}
\usage{
table8.3
}
\description{
The data in Table 8.3 as part of a preliminary study of a possible link between football helmet design and neck injuries.
}
\keyword{datasets}
