% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table3.4.R
\docType{data}
\name{table3.4}
\alias{table3.4}
\title{Table 3.4}
\format{
A dataframe with 10 rows and 4 columns
\describe{
\item{LocationNumber}{ID}
\item{y1}{available soil calcium}
\item{y2}{exchangeable soil calcium}
\item{y3}{turnip green calcium}
}
}
\source{
Kramer and Jensen (1969)
}
\usage{
table3.4
}
\description{
Table 3.4 Calcium in Soil and Turnip Greens
}
\keyword{datasets}
